@@ -1,30 +1,47 @@
 {
-    "name": "react-dnd-chess",
-    "version": "0.1.0",
-    "description": "Simple chessboard app made in React",
-    "engines": {
-        "node": "0.12.7"
-    },
-    "main": "main.js",
-    "directories": {
-        "test": "tests"
-    },
-    "repository": {
-        "type": "git",
-        "url": "ssh://git@github.com:reuben-bradley/react-chessboard.git"
-    },
-    "author": {
-        "name": "Reuben Bradley",
-        "email": "reuben.m.bradley@gmail.com"
-    },
-    "license": "MIT",
-    "dependencies": {
-        "webpack": "^1.12.4",
-        "react": "^0.14.2",
-        "react-dom": "^0.14.2",
-        "react-dnd": "^2.0.2",
-        "react-dnd-html5-backend": "^2.0.0",
-        "jsx-loader": "^0.13.2"
-    }
+  "name": "react-dnd-chess",
+  "version": "0.1.0",
+  "private": true,
+  "description": "Simple chessboard app made in React",
+  "repository": {
+    "type": "git",
+    "url": "ssh://git@github.com:reuben-bradley/react-chessboard.git"
+  },
+  "author": {
+    "name": "Reuben Bradley",
+    "email": "reuben.m.bradley@gmail.com"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test",
+    "eject": "react-scripts eject"
+  },
+  "eslintConfig": {
+    "extends": "react-app"
+  },
+  "license": "MIT",
+  "dependencies": {
+    "@testing-library/jest-dom": "^4.2.4",
+    "@testing-library/react": "^9.5.0",
+    "@testing-library/user-event": "^7.2.1",
+    "prop-types": "^15.7.2",
+    "react": "^16.13.1",
+    "react-dnd": "^11.1.3",
+    "react-dnd-html5-backend": "^11.1.3",
+    "react-dom": "^16.13.1",
+    "react-scripts": "^3.4.3"
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  }
 }
-
