@@ -1,27 +1,28 @@
+import React from 'react';
+import PropTypes from 'prop-types';
+import { DragSource } from 'react-dnd';
 
-var React = require('react');
-var Constants = require('./Constants');
-var DragSource = require('react-dnd').DragSource;
+import Constants from 'Constants';
 
-var ChessPiece = React.createClass({
-    propTypes: {
-        connectDragSource: React.PropTypes.func.isRequired,
-        isDragging: React.PropTypes.bool.isRequired,
-        piece: React.PropTypes.string.isRequired
-    },
-    
-    getDefaultProps: function () {
+class ChessPiece extends React.Component {
+    getDefaultProps() {
         return {
             piece: Constants.PieceStrings.WHITE_PAWN
         };
-    },
-    render: function () {
+    }
+    render() {
         return this.props.connectDragSource(
             <span className="piece" style={{ opacity: this.props.isDragging ? 0.5 : 1 }}>{this.props.piece}</span>
         );
     }
-});
+};
 
-module.exports = DragSource(
+ChessPiece.propTypes = {
+    connectDragSource: PropTypes.func.isRequired,
+    isDragging: PropTypes.bool.isRequired,
+    piece: PropTypes.string.isRequired
+};
+
+export default DragSource(
     Constants.ItemTypes.PIECE, Constants.DragSources.PIECE, Constants.DragSources.collect
 )(ChessPiece);
